server/requirements.txt
fastapi
uvicorn
bs4
requests
aiohttp
urljoin
protobuf==5.29.3
grpcio
asyncio
python-multipart
python-jose
passlib
bcrypt
